+++ Games/EchoGame/src/main/java/jokatu/game/games/echo/EchoConfiguration.java
@@ -2,6 +2,7 @@ package jokatu.game.games.echo;
 
 import jokatu.components.GameComponent;
 import jokatu.components.config.GameConfiguration;
+import jokatu.game.games.echo.game.EchoGameFactory;
 import jokatu.game.games.echo.views.EchoViewResolverFactory;
 import jokatu.game.input.InputDeserialiser;
 import ophelia.collections.BaseCollection;
@@ -9,7 +10,7 @@ import ophelia.collections.set.Singleton;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import static jokatu.game.games.echo.EchoGame.ECHO;
+import static jokatu.game.games.echo.game.EchoGame.ECHO;
 
 /**
  * @author steven
