+++ Games/EchoGame/src/main/java/jokatu/game/games/echo/game/Echo.java
@@ -1,4 +1,4 @@
-package jokatu.game.games.echo;
+package jokatu.game.games.echo.game;
 
 import jokatu.game.event.PublicGameEvent;
 import jokatu.game.games.echo.input.EchoInput;
