+++ Games/EchoGame/src/main/java/jokatu/game/games/echo/game/EchoGameFactory.java
@@ -1,4 +1,4 @@
-package jokatu.game.games.echo;
+package jokatu.game.games.echo.game;
 
 import jokatu.game.AbstractGameFactory;
 import jokatu.game.GameID;
